test(frontend): split chat rendering coverage
